Controller Reflectiveness Improved
Played around with the roughness, should be fine now. The material is
still fine, tested in another project. Seems to be that it is getting a
lot of added reflecdtiveness. Best way to fix this would be changing
whatever is adding more reflectiveness to the material but for now this
fix works, even if it makes the controller lose almsot all
reflectiveness.